修改收入报表

+++ resto-shop-web/src/main/java/com/resto/shop/web/controller/business/OrderExceptionController.java
@@ -61,7 +61,9 @@
          //查寻orderpaymentItem中result_data 为 {}
         OrderPaymentItem oi =  orderPaymentItemService.selectByOrderIdAndResultData(orderId);
          //删除该订单项
-         orderPaymentItemService.delete(oi.getId());
+         if(null!=oi.getId()){
+             orderPaymentItemService.delete(oi.getId());
+         }
          orderService.update(o);
          orderService.cancelExceptionOrder(orderId);
          return Result.getSuccess();
